feat: Add index page when user hasn't onboard +fixes
Added an index page for when a user hasn't explicitely onboarded to the
app. This is in preparation for a proper onboarding workflow. If the
user hasn't onboarded yet, they will then be prompted to choose a name
before being redirected to the app. A number of bugs were also
discovered during implementation, such as: Web worker being run multiple
times through it's initialization, generating multiple DIDs. This caused
issues with the server listings showing our own DID as a server
automatically. Some tweaks to the UI have also been made to show the DID
in a nicer box display in the chat list. We'll want to put that under
the developer settings and somewhere else in the future.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>